first thr version done/not tested

+++ tools/socktap/thr_s_application.hpp
@@ -14,7 +14,7 @@ public:
 
 private:
 
-    unsigned int counter = 0;
+    uint16_t counter = 0;
 
     void schedule_timer();
     void message_counter();
